+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [Added] Env column for Windows (environment variables)
 * [Added] RecvBytes and SendBytes columns for Windows (network I/O rate; needs Windows 11 or later)
 * [Added] WorkDir column for Windows (current working directory)
+* [Added] WorkDir column for macOS (current working directory)
 * [Added] `--thread` support for Windows
 * [Changed] Group and Gid columns for Windows read the process token only when the column is displayed
 * [Fixed] ReadBytes / WriteBytes divided by a mis-scaled interval (seconds added to milliseconds)